update api response types

+++ pages/api/deleteResumeById.ts
@@ -1,29 +1,58 @@
 import { NextApiRequest, NextApiResponse } from "next";
+import { unstable_getServerSession } from "next-auth";
+import { authOptions } from "./auth/[...nextauth]";
 import FileModel from "../../models/FileModel.model";
 import dbConnect from "../../utils/database";
-import { ResponseSuccess, ResponseError } from "../../custom2.d";
+import { ApiResponse } from "../../custom2.d";
 
 const handler = async (
     req: NextApiRequest,
-    res: NextApiResponse<ResponseSuccess | ResponseError>
+    res: NextApiResponse<ApiResponse>
 ) => {
     try {
         switch (req.method) {
             case "POST": {
-                await dbConnect();
-                const fileId = JSON.parse(req.body).fileId;
-                if (fileId) {
-                    const x = await FileModel.deleteOne({ id: fileId });
-                    console.log(x);
+                const session = await unstable_getServerSession(
+                    req,
+                    res,
+                    authOptions
+                );
+
+                if (session) {
+                    await dbConnect();
+                    const fileId = JSON.parse(req.body).fileId;
+                    if (fileId) {
+                        const resp = await FileModel.deleteOne({ id: fileId });
+                        if (resp.deletedCount > 0)
+                            res.status(200).json({
+                                data: { message: "Successfully delete file." },
+                            });
+                    } else {
+                        res.status(404).json({
+                            error: {
+                                code: 404,
+                                message: "Failed to delete file.",
+                            },
+                        });
+                    }
+                } else {
+                    res.status(401).json({
+                        error: { code: 401, message: "Unauthorized" },
+                    });
                 }
-                res.status(200).json({ data: {} });
+
                 break;
             }
             default:
-                throw new Error(`${req.method} is not allowed.`);
+                res.status(405).json({
+                    error: { code: 405, message: `${req.method} not allowed.` },
+                });
         }
     } catch (ex) {
-        res.status(500).json({ error: { code: 500, message: ex } });
+        console.log(ex);
+        res.status(500).json({
+            error: { code: 500, message: "Internal Server Error" },
+        });
     }
 };
